% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{as.eigen.hky}
\alias{as.eigen.hky}
\title{K80 eigen decomposition}
\usage{
\method{as.eigen}{hky}(hky.rates, mc.stat, scale = F)
}
\arguments{
\item{hky.rates}{transition and transversion rates}

\item{mc.stat}{stationary distribution}

\item{scale}{if TRUE then scale the transition matrix. Defaults to FALSE.}
}
\description{
This function defines the K80 model via its eigen decomposition. I'm not sure why it won't show up.
}
\examples{
as.eigen.hky(hky.rates, mc.stat, scale = F)
}
\keyword{OLS}
\keyword{phylogeny,}
