% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{gamma.ls.loss}
\alias{gamma.ls.loss}
\title{Loss function for gamma-distributed rate heterogeneity}
\usage{
\method{gamma}{ls.loss}(log.par, my.topology, seq.table)
}
\arguments{
\item{log.par}{natural log of the branch lengths and transition/transversion ratio}

\item{my.topology}{the tree on which to calculate the loss function}

\item{seq.table}{a nucleotide sequence alignment}
}
\description{
This function calculates the loss function value for a JC69+G model
}
\examples{
gamma.ls.loss(log.par, my.topology, seq.table)
}
\keyword{OLS}
\keyword{phylogeny,}
