% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{new.ls.fit.K80}
\alias{new.ls.fit.K80}
\title{Least Squares K80 without distances}
\usage{
new.ls.fit.K80(
  my.topology,
  seq.table,
  init.brlen = NULL,
  init.kappa = NULL,
  method = "nlminb",
  low = -100,
  high = NULL,
  high.k = 100,
  rel.tol = 1e-04,
  starttests = TRUE,
  kkt = TRUE
)
}
\arguments{
\item{my.topology}{the tree on which to calculate the loss function}

\item{seq.table}{a nucleotide sequence alignment}

\item{log.br.len}{natural log of the branch lengths}

\item{log.kappa}{natural log of the transition/transversion ratio}
}
\description{
This function calculates the distance free loss function under the Kimura 2 parameter model (K80).
}
\examples{
new.loss.K80(log.br.len, kappa, my.topology, seq.table)
}
\keyword{OLS}
\keyword{phylogeny,}
