% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{new.ls.loss}
\alias{new.ls.loss}
\title{L2 loss function}
\usage{
new.ls.loss(log.branch.length, my.topology, seq.table)
}
\arguments{
\item{log.branch.length}{natural log of the branch lengths of the phylogeny}

\item{my.topology}{a phylogeny in ape format}

\item{seq.table}{A numeric matrix of values 1-4 for each nucleotide type. Use function \code{read.phylosim.nuc} to transform character matrix to numeric if needed.}
}
\description{
This function calculates the L2 loss as described in...
}
\examples{
phylo.ML(seq.table)
}
\keyword{OLS}
\keyword{phylogeny,}
