% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{new.phylo.ls}
\alias{new.phylo.ls}
\title{A phyloLSnoDist Function}
\usage{
new.phylo.ls(
  seq.table,
  initvals = NULL,
  search.all = FALSE,
  method = "nlminb",
  low = -100,
  high = 2,
  tol = 1e-10
)
}
\arguments{
\item{seq.table}{a nucleotide sequence alignment, formatted as an n x s character matrix where n = # of taxa, s = # of sites}

\item{initvals}{starting values for each branch}

\item{search.all}{if TRUE, an exhaustive search across all topologies will be performed. Otherwise, an NNI search will be performed.}

\item{method}{optimization method for optimx function}

\item{low}{lower bound for numeric optimization}

\item{high}{upper bound for numeric optimization}

\item{tol}{in NNI search, keep searching if improvement is at least this amount}
}
\value{
An unrooted phylogeny
}
\description{
This function performs phylogenetic inference via least squares with our new loss. It allows for an exhaustive search among
all possible topologies (if not, it will do an NNI search, starting from the NJ tree). This function infers an unrooted tree.
}
\examples{
new.phylo.ls(seq.table)
}
\keyword{OLS}
\keyword{phylogeny,}
