% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{phylo.ls.nodist}
\alias{phylo.ls.nodist}
\title{Least squares phylogenetic inference without distances}
\usage{
phylo.ls.nodist(
  alignment,
  initvals = NULL,
  search.all = FALSE,
  method = "nlminb",
  ts = FALSE,
  low = -100,
  high = NULL,
  rel.tol = 1e-04,
  bound_rm = TRUE
)
}
\arguments{
\item{alignment}{a nucleotide sequence alignment, of class \code{phyDat}. Use \code{phyDat}
function from \code{phangorn} package to transform data if necessary.}

\item{initvals}{starting values for each branch}

\item{search.all}{if TRUE, an exhaustive search across all topologies will be performed. Otherwise, an NNI search will be performed.}

\item{method}{optimization method for optimx function}

\item{low}{lower bound for numeric optimization}

\item{high}{upper bound for numeric optimization}

\item{tol}{in NNI search, keep searching if improvement is at least this amount}
}
\value{
An unrooted phylogeny
}
\description{
This function performs phylogenetic inference via least squares with our new loss function. It allows for an exhaustive search among
all possible topologies (if not, it will do an NNI search, starting from the NJ tree). This function infers an unrooted tree.
}
\examples{
phylo.ls.nodist(seq.table)

}
\keyword{OLS}
\keyword{phylogeny,}
