% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{phylo.ls}
\alias{phylo.ls}
\title{Least squares phylogenetic inference}
\usage{
phylo.ls(
  alignment,
  set.neg.to.zero = TRUE,
  search.all = FALSE,
  model = "JC69",
  tol = 1e-10
)
}
\arguments{
\item{set.neg.to.zero}{if TRUE, negative branch lengths will be converted to 0}

\item{search.all}{if TRUE, an exhaustive search across all topologies will be performed. Otherwise, an NNI search will be performed.}

\item{model}{substitution model for which to calculate the distance matrix}

\item{tol}{in NNI search, keep searching if improvement is at least this amount}

\item{seq.table}{a nucleotide sequence alignment, formatted as an n x s character matrix where n = # of taxa, s = # of sites}
}
\description{
This function performs phylogenetic inference via ordinary least squares. It was written taking elements
from Liam Revell's optim.phylo.ls. The main difference is that it allows for an exhaustive search among
all possible topologies (if not, it will do an NNI search, starting from the NJ tree). This function infers an unrooted tree.
}
\examples{
phylo.ls(seq.table)
}
\keyword{OLS}
\keyword{phylogeny,}
