% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{read.phylosim.nuc}
\alias{read.phylosim.nuc}
\title{Maximum likelihood phylogenetic inference}
\usage{
read.phylosim.nuc(alignment)
}
\arguments{
\item{alignment}{a nucleotide sequence alignment, of class phyDat}

\item{search.all}{if TRUE, an exhaustive search across all topologies will be performed. Otherwise, an NNI search will be performed.}

\item{tol}{in NNI search, keep searching if improvement is at least this amount. Used 1e-8 as default value consistent with phangorn.}
}
\value{
An unrooted phylogeny
}
\description{
This is basically a wrapper around the pml function from the phangorn package, made to suit our purposes here
and perform either an exhaustive search through all topologies or an NNI search with pml as the workhorse.
I don't really recommend that anyone should use this function instead of optim.pml from phangorn unless you see a strong reason to.
}
\examples{
phylo.ML(seq.table)
}
\keyword{OLS}
\keyword{phylogeny,}
