% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{regular.ls.fit}
\alias{regular.ls.fit}
\title{A phyloLSnoDist function}
\usage{
regular.ls.fit(init.brlen, my.topology, seq.dist)
}
\arguments{
\item{init.brlen}{Initial branch lengths at which to start the search}

\item{my.topology}{a phylogeny in ape format}

\item{seq.dist}{not sure}
}
\description{
Not sure if I need this one anymore either.
}
\examples{
regular.ls.fit(init.brlen, my.topology, seq.dist)
}
\keyword{OLS}
\keyword{phylogeny,}
