% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{regular.ls.loss}
\alias{regular.ls.loss}
\title{A phyloLSnoDist function}
\usage{
regular.ls.loss(log.branch.length, my.topology, seq.dist)
}
\arguments{
\item{log.branch.length}{Natural log of the branch lengths at which to calculate the loss}

\item{my.topology}{A phylogeny in ape format}

\item{seq.dist}{not sure}
}
\description{
This might not be needed. Revisit later.
}
\examples{
regular.ls.loss(log.branch.length, my.topology, seq.dist)
}
\keyword{OLS}
\keyword{phylogeny,}
