% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{new.ls.fit.optimx}
\alias{new.ls.fit.optimx}
\title{Find optimal branch lengths according to L2 loss}
\usage{
new.ls.fit.optimx(
  my.topology,
  seq.table,
  init.brlen = NULL,
  method = "nlminb",
  ts = FALSE,
  low = -100,
  high = NULL,
  max.attempts = 5,
  starttests = TRUE,
  kkt = TRUE,
  rel.tol = 1e-04
)
}
\arguments{
\item{my.topology}{a phylogenetic topology on which to optimize the branch lengths}

\item{seq.table}{a nucleotide sequence alignment, of class \code{phyDat}. Use \code{phyDat}
function from \code{phangorn} package to transform data if necessary.}

\item{init.brlen}{initial branch lengths to use in the optimization routine}

\item{method}{option for optimx function to choose which optimization method to use}

\item{ts}{option to use transversion distance only}

\item{low}{contraint on optimization. Defaults to -100 which will constrain br len to be min of 3.72e-44}

\item{high}{constraint on optimization. Defaults to 2 which will constrain br len to be max of e^2 ~ 7.389}

\item{starttests}{control parameter for optimx, default to \code{FALSE} for speed}

\item{kkt}{control parameter for optimx, default to \code{FALSE} for speed}

\item{initvals}{starting values for each branch}
}
\value{
An unrooted phylogeny
}
\description{
This function takes a topology and nucleotide sequence alignment, and finds the optimal branch lengths
according to the new loss function
}
\examples{
phylo.ML(seq.table)
}
\keyword{OLS}
\keyword{phylogeny,}
