% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ls_slim.R
\name{new.ls.loss}
\alias{new.ls.loss}
\title{L2 loss function}
\usage{
new.ls.loss(log.branch.length, my.topology, seq.table, regist.matrix)
}
\arguments{
\item{log.branch.length}{natural log of the branch lengths of the phylogeny}

\item{my.topology}{a phylogeny in ape format}

\item{regist.matrix}{4/6/20 I don't know why this is an argument if it is defined in the function...}

\item{seq.dist}{4/6/20 not needed?}
}
\description{
This function calculates the L2 loss as described in...
}
\examples{
phylo.ML(seq.table)
}
\keyword{OLS}
\keyword{phylogeny,}
